+++ server/apps/trak/views/__init__.py
@@ -1,4 +1,4 @@
 from .handler import HandlerSearch, HandlerView
 from .manifest import ManifestView
-from .profile import RcraProfileView, SyncProfile
+from .rcra_profile import RcraProfileView, SyncProfile
 from .sites import SiteApi, SiteList, SiteManifest
