+++ server/requirements.txt
@@ -6,7 +6,7 @@ django-celery-results==2.4.0
 djangorestframework==3.14.0
 django-cors-headers==3.13.0
 gunicorn==20.1.0
-emanifest==2.0.6
+emanifest==2.0.7
 idna==3.4
 psycopg2-binary==2.9.5
 pytz==2022.7
